+++ glade_forms/lib/src/model/glade_model.dart
@@ -6,15 +6,26 @@ abstract class GladeModel extends ChangeNotifier {
   List<GladeInput<Object?>> _lastUpdates = [];
   bool _groupEdit = false;
 
+  /// Returns true if all inputs are valid.
   bool get isValid => inputs.every((input) => input.isValid);
 
+  /// Returns true if any input is not valid.
   bool get isNotValid => !isValid;
 
+  /// Returns true if all inputs are pure.
+  ///
+  /// Input is pure if its value is same as initial value and value was never updated.
+  ///
+  /// Pure can be reset when [setInputValuesAsNewInitialValues] or [resetToInitialValue] on model or its inputs are called.
   bool get isPure => inputs.every((input) => input.isPure);
 
-  bool get isUnchanged => inputs.where((input) => input.trackUnchanged).every((input) => input.isUnchanged);
-
+  /// Returns true if model is not pure.
   bool get isDirty => !isPure;
+
+  /// Returns true if all inputs are unchanged.
+  ///
+  /// Input is unchanged if its value is same as initial value, even if value was updated into initial value.
+  bool get isUnchanged => inputs.where((input) => input.trackUnchanged).every((input) => input.isUnchanged);
 
   ErrorTranslator<Object?> get defaultErrorTranslate => (error, key, devMessage, dependencies) => devMessage;
 
@@ -121,6 +132,7 @@ abstract class GladeModel extends ChangeNotifier {
     notifyListeners();
   }
 
+  /// Notifies dependant inputs about changes.
   void notifyDependencies() {
     final updatedKeys = _lastUpdates.map((e) => e.inputKey).toSet();
     for (final input in inputs) {
@@ -131,28 +143,27 @@ abstract class GladeModel extends ChangeNotifier {
     }
   }
 
-  /// Sets a new pure state for all inputs in the model.
+  /// Sets the initial values of all inputs to their current values.
   ///
-  /// When [copyValueToInitialValue] is true, input's initialValue is overriden by current value.
-  void setAsNewPure({
-    bool invokeUpdate = true,
-    bool copyValueToInitialValue = false,
-  }) {
+  /// [shouldTriggerOnChange] - if true, onChange callbacks will be triggered.
+  void setInputValuesAsNewInitialValues({bool shouldTriggerOnChange = true}) {
     for (final input in inputs) {
-      input.setAsNewPure(
-        invokeUpdate: invokeUpdate,
-        copyValueToInitialValue: copyValueToInitialValue,
+      input.setNewInitialValue(
+        initialValue: () => input.value,
+        shouldResetToInitialValue: true,
+        shouldTriggerOnChange: shouldTriggerOnChange,
       );
     }
-    if (invokeUpdate) {
-      notifyListeners();
-    }
+
+    notifyListeners();
   }
 
-  /// Resets all inputs in the model to their initial values and sets them as pure.
-  void resetToPure() {
+  /// Resets all inputs in the model to their initial values.
+  ///
+  /// [shouldTriggerOnChange] - if true, onChange callbacks will be triggered.
+  void resetToInitialValue({bool shouldTriggerOnChange = true}) {
     for (final input in inputs) {
-      input.resetToPure();
+      input.resetToInitialValue(shouldTriggerOnChange: shouldTriggerOnChange);
     }
     notifyListeners();
   }